// src/interfaces/themes.interface.ts
export interface Theme {
  id: string;
  title: string;
  dark: boolean;
  background_default: string;
  background_paper: string;
  text_primary: string;
  text_secondary: string;
  text_disabled: string;
  text_hint: string;
  primary_main: string;
  primary_light: string;
  primary_dark: string;
  primary_contrast: string;
  secondary_main: string;
  secondary_light: string;
  secondary_dark: string;
  secondary_contrast: string;
  error_main: string;
  error_light: string;
  error_dark: string;
  error_contrast: string;
  success_main: string;
  success_light: string;
  success_dark: string;
  success_contrast: string;
}